pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "git-remote-get"
version = "0.0.1"
description = ""
authors = ["01Joseph-Hwang10 <joseph95501@gmail.com>"]
license = "MIT"
readme = "README.md"
exclude = ["**/tests", "**/package.json", "**/project.json"]
[[tool.poetry.packages]]
include = "gget"

[tool.black]
line-length = 88
target-version = ["py311"]
include = "\\.pyi?$"
extend-exclude = "/(\n  # The following are specific to Black, you probably don't want those.\n  tests/data\n  | profiling\n)/\n"

[tool.poetry.scripts]
gget = "gget.__main__:cli"

[tool.poe.tasks]
test = "python -m pytest"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
python-dotenv = "^1.0.0"
click = "^8.1.7"
pygithub = "^2.2.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.9.1"
pre-commit = "^3.4.0"
commitizen = "^3.10.0"
poethepoet = "^0.24.4"

[tool.poetry.group.test.dependencies]
faker = "^20.1.0"
pytest = "^7.4.3"
pytest-order = "^1.2.0"
